move components to components
